+++ vstools/utils/other.py
@@ -6,8 +6,10 @@ from typing import TYPE_CHECKING
 
 __all__ = [
     'IS_DOCS',
-    'get_nvidia_version'
+    'get_nvidia_version',
+    'is_gpu_available'
 ]
+
 
 IS_DOCS = not TYPE_CHECKING and (
     # if loaded from sphinx
@@ -44,3 +46,25 @@ def get_nvidia_version() -> tuple[int, int] | None:
             return _str_to_ver(smi.stdout.splitlines()[5].decode().split(':')[-1])
 
     return None
+
+
+def is_gpu_available() -> bool:
+    """Check if any GPU is available."""
+
+    try:
+        smi = run(['nvidia-smi'], capture_output=True, text=True)
+    except FileNotFoundError:
+        pass
+    else:
+        if smi.returncode == 0:
+            return True
+
+    try:
+        rocm_smi = run(['rocm-smi'], capture_output=True, text=True)
+    except FileNotFoundError:
+        pass
+    else:
+        if rocm_smi.returncode == 0:
+            return True
+
+    return False